+++ h2integrate/storage/battery/battery_baseclass.py
@@ -37,50 +37,8 @@ class BatteryPerformanceBaseClass(om.ExplicitComponent):
             val=0.0,
             copy_shape="electricity_in",
             units="kW/h",
-            desc="Electricty output from Battery only"
+            desc="Electricty output from Battery only",
         )
-
-    def compute(self, inputs, outputs):
-        """
-        Computation for the OM component.
-
-        For a template class this is not implement and raises an error.
-        """
-
-        raise NotImplementedError("This method should be implemented in a subclass.")
-
-
-class BatteryCostBaseClass(om.ExplicitComponent):
-    def initialize(self):
-        self.options.declare("driver_config", types=dict)
-        self.options.declare("plant_config", types=dict)
-        self.options.declare("tech_config", types=dict)
-
-    def setup(self):
-        # Define outputs: CapEx and OpEx costs
-        self.add_output("CapEx", val=0.0, units="USD", desc="Capital expenditure")
-        self.add_output("OpEx", val=0.0, units="USD/year", desc="Operational expenditure")
-
-    def compute(self, inputs, outputs):
-        """
-        Computation for the OM component.
-
-        For a template class this is not implement and raises an error.
-        """
-
-        raise NotImplementedError("This method should be implemented in a subclass.")
-
-
-class BatteryFinanceBaseClass(om.ExplicitComponent):
-    def initialize(self):
-        self.options.declare("driver_config", types=dict)
-        self.options.declare("plant_config", types=dict)
-        self.options.declare("tech_config", types=dict)
-
-    def setup(self):
-        self.add_input("CapEx", val=0.0, units="USD")
-        self.add_input("OpEx", val=0.0, units="USD/year")
-        self.add_output("NPV", val=0.0, units="USD", desc="Net present value")
 
     def compute(self, inputs, outputs):
         """