+++ tests/h2integrate/test_hydrogen/test_geologic_h2.py
@@ -17,7 +17,7 @@ def test_natural_geoh2(subtests):
         h2_prod = h2i_nat.plant.geoh2.natural_geoh2_performance.get_val("hydrogen_out")
         assert np.mean(h2_prod) == approx(48.94393478, 1e-6)
     with subtests.test("LCOH"):
-        lcoh = h2i_nat.plant.geoh2.geoh2_financial.get_val("LCOH")
+        lcoh = h2i_nat.plant.geoh2.geoh2_finance.get_val("LCOH")
         assert lcoh == approx(3.14353262, 1e-6)
 
 
@@ -29,5 +29,5 @@ def test_stimulated_geoh2(subtests):
         h2_prod = h2i_stim.plant.geoh2.stimulated_geoh2_performance.get_val("hydrogen_out")
         assert np.mean(h2_prod) == approx(108.49331847, 1e-6)
     with subtests.test("LCOH"):
-        lcoh = h2i_stim.plant.geoh2.geoh2_financial.get_val("LCOH")
+        lcoh = h2i_stim.plant.geoh2.geoh2_finance.get_val("LCOH")
         assert lcoh == approx(1.8666849, 1e-6)